+++ src/Components/SearchModal/index.tsx
@@ -1,134 +1,144 @@
-import axios from 'axios';
-import { Fragment, useEffect, useRef, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { useLocation, useNavigate } from 'react-router-dom';
-
-interface SearchListType {
-  id: number;
-  postedAt: string;
-  titleSnippet: string;
-  contentSnippet: string;
-}
+import { useAppDispatch, useAppSelector } from 'hooks';
+import React, { Fragment, useEffect, useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
+import { search, searchModalActions } from 'redux/slice/searchModal-slice';
 
 export const SearchModal = () => {
-  const [searchValue, setSearchValue] = useState('');
-  const [searchList, setSearchList] = useState<SearchListType[]>([]);
-  const [loading, setLoading] = useState(true);
-
-  const BACK_URL = process.env.REACT_APP_BACK_URL;
-  const BACK_PORT = process.env.REACT_APP_BACK_DEFAULT_PORT;
-
-  const location = useLocation();
-  const pathname = location.pathname;
-
-  const SearchDivRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const closeClickOutside = (e: any) => {
-      if (
-        SearchDivRef.current &&
-        !SearchDivRef.current.contains(e.target as Node)
-      ) {
-        setSearchValue('');
-      }
-    };
-    document.addEventListener('click', closeClickOutside);
-    return () => {
-      document.removeEventListener('click', closeClickOutside);
-    };
-  }, [SearchDivRef]);
-
-  useEffect(() => {
-    if (searchValue.trim() !== '') {
-      const timer = setTimeout(() => {
-        axios
-          .get<SearchListType[]>(
-            `${BACK_URL}:${BACK_PORT}/search?query=${searchValue}`,
-            { timeout: 5000 }
-          )
-          .then(response => {
-            setSearchList(response.data);
-            setLoading(false);
-          });
-      }, 500);
-      return () => clearTimeout(timer);
-    }
-    if (searchValue.trim() === '') {
-      setSearchList([]);
-      setLoading(true);
-    }
-  }, [searchValue]);
-
-  const showResult = (searchList: SearchListType[]) => {
-    if (!loading && searchList.length !== 0) {
-      return searchList.map(result => {
-        return (
-          <Link key={result.id} to={'/feed/' + result.id}>
-            <div>
-              <div className="flex justify-between items-center mb-4">
-                <div className="w-32 font-bold">
-                  {result.titleSnippet ? result.titleSnippet : '-'}
-                </div>
-                <div className="w-24 text-sm text-buttongray">
-                  {result.postedAt && result.postedAt.slice(0, -8)}
-                </div>
-              </div>
-              <div className="w-48 min-h-[2.5rem]">
-                {result.contentSnippet ? result.contentSnippet : '-'}
-              </div>
-            </div>
-          </Link>
-        );
-      });
-    }
-    if (!loading && searchList.length === 0) {
-      return <div className="text-center">검색 결과가 없습니다😥</div>;
-    }
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(false);
+  const dispatch = useAppDispatch();
+  const handleModal = searchModalActions.handleModal;
+  const isSearchModalOpen = useAppSelector(state => state.search.isModalOpen);
+  const setSearchWord = (e: React.ChangeEvent<HTMLInputElement>) => {
+    dispatch(searchModalActions.getSearchWord(e.target.value));
   };
+  const searchData = useAppSelector(state => state.search);
 
-  const getSearchValue = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchValue(e.target.value);
+  useEffect(() => {
+    if (searchData.searchWord.trim() !== '') {
+      try {
+        setLoading(true);
+        const timer = setTimeout(() => {
+          dispatch(search(searchData.searchWord));
+          setLoading(false);
+        }, 500);
+        return () => {
+          setLoading(false);
+          clearTimeout(timer);
+        };
+      } catch (error) {
+        setLoading(false);
+        setError(true);
+      }
+    }
+  }, [searchData.searchWord, dispatch]);
+
+  useEffect(() => {
+    if (searchData.searchWord.trim() === '' || isSearchModalOpen === false) {
+      dispatch(searchModalActions.resetSearchModal());
+      dispatch(searchModalActions.getSearchWord(''));
+    }
+  }, [searchData.searchWord, isSearchModalOpen, dispatch]);
+
+  const searchResult = () => {
+    if (searchData.searchResult?.length === 0 && !loading) {
+      return (
+        <div className="justify-selt-center self-center">
+          검색 결과가 없습니다😥
+        </div>
+      );
+    }
+
+    if (error) {
+      return <div>잠시 후 다시 시도해주세요.</div>;
+    }
   };
 
   const navigate = useNavigate();
 
-  const search = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (searchValue.trim() !== '' && e.key === 'Enter') {
-      let url = `/search?query=${searchValue}`;
+  const goToSearchPage = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (searchData.searchWord.trim() !== '' && e.key === 'Enter') {
+      let url = `/search?query=${searchData.searchWord}`;
+      dispatch(searchModalActions.handleModal());
       navigate(url);
     }
   };
 
+  const highlightedText = (text: string) => {
+    const strs = text.split(new RegExp(`(${searchData.searchWord})`, 'gi'));
+    return (
+      <>
+        {strs.map((str, idx) =>
+          str.toLowerCase() === searchData.searchWord.toLowerCase() ? (
+            <span key={idx} className="text-mainblue underline">
+              {str}
+            </span>
+          ) : (
+            str
+          )
+        )}
+      </>
+    );
+  };
+
   return (
     <Fragment>
-      {searchValue.trim() !== '' && (
-        <div className="fixed w-full h-screen bg-black/[.1] top-1/2 left-1/2 -translate-y-1/2 -translate-x-1/2 overflow-hidden z-50" />
-      )}
-      {pathname !== '/search' && (
-        <div ref={SearchDivRef}>
-          <input
-            className="w-80 mr-4 p-2.5 border border-buttongray rounded-md focus:outline-none z-50"
-            type="search"
-            placeholder="검색어를 입력해주세요"
-            onChange={e => getSearchValue(e)}
-            onKeyUp={search}
-          />
-          {searchValue && (
-            <div className="absolute w-80 p-4 bg-white border border-buttongray rounded-md z-50">
-              <div className="flex flex-col w-full max-h-96 p-4 gap-4 bg-white overflow-y-auto z-50">
-                {loading ? (
-                  <div className="text-center">Loading...</div>
-                ) : (
-                  showResult(searchList)
-                )}
-              </div>
-              <Link to={'/search?query=' + searchValue}>
-                <div className="w-full h-4 mt-1.5 text-center cursor-pointer">
-                  더보기
-                </div>
-              </Link>
+      {isSearchModalOpen && (
+        <div className="hidden md:block">
+          <div className="fixed w-full h-screen bg-black/[.1] top-1/2 left-1/2 -translate-y-1/2 -translate-x-1/2 overflow-hidden z-50" />
+          <div className="absolute top-40 m-auto left-0 right-0 z-50 w-2/5 h-1/2 bg-white flex flex-col rounded-md py-8 px-12 animate-search-open">
+            <div className="flexCenterAlign w-full mt-4">
+              <div className="w-4 h-4 mr-2 bg-[url('./assets/images/search.png')] bg-no-repeat bg-cover" />
+              <input
+                className="w-full mr-4 p-2.5 border-b focus:outline-none z-50"
+                type="search"
+                placeholder="검색어를 입력해주세요"
+                onChange={e => setSearchWord(e)}
+                onKeyUp={goToSearchPage}
+              />
             </div>
-          )}
+            <div
+              className=" absolute w-4 h-4 mr-2 bg-[url('./assets/images/close.png')] bg-no-repeat bg-cover cursor-pointer top-8 right-8"
+              onClick={() => {
+                dispatch(handleModal());
+              }}
+            />
+            <div className="w-full h-80 bg-white p-4">
+              <div className="flex flex-col items-start w-full h-full p-4 gap-4 bg-white overflow-y-auto z-50 border-b">
+                {loading && (
+                  <div className="w-8 h-8 justify-selt-center self-center bg-[url('./assets/images/loader.png')] bg-no-repeat bg-cover animate-spin" />
+                )}
+                {searchData.searchResult?.map(item => (
+                  <Link key={item.id} to={`/feed/${item.id}`}>
+                    <div
+                      className="flex flex-col gap-2"
+                      onClick={() => dispatch(searchModalActions.handleModal())}
+                    >
+                      <div className="flex gap-4">
+                        <span className="font-bold">
+                          {highlightedText(item.titleSnippet)}
+                        </span>
+                        <span className="text-sm text-buttongray">
+                          {item.postedAt.slice(0, 10)}
+                        </span>
+                      </div>
+                      <span>{highlightedText(item.contentSnippet)}</span>
+                    </div>
+                  </Link>
+                ))}
+                {searchResult()}
+              </div>
+              {searchData.searchResult?.length !== 0 && (
+                <Link
+                  className="flexCenterAlign absolute bottom-8 right-8 text-center cursor-pointer"
+                  to={`/search?query=${searchData.searchWord}`}
+                >
+                  <span>더보기</span>
+                  <div className="w-4 h-4 ml-2 bg-[url('./assets/images/toggleDown.png')] rotate-180 bg-no-repeat bg-cover" />
+                </Link>
+              )}
+            </div>
+          </div>
         </div>
       )}
     </Fragment>
