+++ src/Components/index.tsx
@@ -9,3 +9,4 @@ export * from './WriteFeed';
 export * from './MyFeeds';
 export * from './MyComments';
 export * from './CheckPassword';
+export * from './SearchModal';
